trackmania_rl/requirements.txt
gym
numpy
pyyaml
torch
pygame